requirements.txt
# amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
billiard==4.2.4
camel-converter==5.0.0
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-parler==2.3
django-prometheus==2.4.1
django-redis==5.4.0
django-simple-history==3.11.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
googleapis-common-protos==1.72.0
grpcio==1.76.0
gunicorn==21.2.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
kombu==5.6.2
meilisearch==0.40.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-celery==0.60b1
opentelemetry-instrumentation-dbapi==0.60b1
opentelemetry-instrumentation-django==0.60b1
opentelemetry-instrumentation-psycopg2==0.60b1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-wsgi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
pillow==12.1.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.14
wrapt==1.17.3
zipp==3.23.0
